add random downsampling with seqtk

+++ .github/scripts/run_illumina_pipeline.sh
@@ -1,4 +1,4 @@
-#!/bin/env/usr bash
+#!/bin/bash
 set -eo pipefail
 
 # Create a Cache Dir
@@ -31,6 +31,54 @@ fi
 
 # Reset and Track
 mv .nextflow.log artifacts/illumina.nextflow.log
+rm -rf results work/ .nextflow*
+
+### Run Illumina Pipeline with Downsampling ###
+nextflow run ./main.nf \
+    -profile conda,test \
+    --cache ./conda_cache_dir \
+    --illumina \
+    --directory $PWD/.github/data/illumina/ \
+    --human_ref $PWD/.github/data/partial_hg38_ref.fa \
+    --downsample \
+    --downsample_count 100 \
+    --downsample_seed 42
+
+### Check Pipeline Outputs ###
+# 1. Num Human Reads
+READS=`awk -F, '$1 == "illumina-18_S15" {print $2}' ./results/removal_summary.csv`
+if [[ "$READS" != "2" ]]; then 
+    echo "Incorrect output: Number of human reads found"
+    echo "  Expected: 2, Got: $READS"
+    exit 1
+fi
+
+# 2. Total Reads Kept
+READS=`awk -F, '$1 == "illumina-4_S12" {print $4}' ./results/removal_summary.csv`
+if [[ "$READS" != "100" ]]; then 
+    echo "Incorrect output: Number of human reads found"
+    echo "  Expected: 100, Got: $READS"
+    exit 1
+fi
+
+# 3. Downsample Maximum
+READS=`awk -F, '$1 == "illumina-4_S12" {print $6}' ./results/removal_summary.csv`
+if [[ "$READS" != "100" ]]; then 
+    echo "Incorrect output: Number of human reads found"
+    echo "  Expected: 100, Got: $READS"
+    exit 1
+fi
+
+# 4. Downsample Seed
+READS=`awk -F, '$1 == "illumina-4_S12" {print $7}' ./results/removal_summary.csv`
+if [[ "$READS" != "42" ]]; then 
+    echo "Incorrect output: Number of human reads found"
+    echo "  Expected: 42, Got: $READS"
+    exit 1
+fi
+
+# Reset and Track
+mv .nextflow.log artifacts/illumina_downsampled.nextflow.log
 rm -rf results work/ .nextflow*
 
 echo "Done"
